refactor(agent-runtime): inline pi-runtime-state into pi-runtime
Exactly two importers remained: pi-runtime.ts used findRuntimeSessionForLookup
and piStatusFromEvents, and http/handlers.ts imported isAgentSettledEvent via
the re-export shim while goal-driver.ts already imported it from shared.

- Inline piStatusFromEvents into the PiSdkSession status getter, dropping the
  eventLog parameter it never read.
- Move findRuntimeSessionForLookup into pi-runtime.ts, concretely typed against
  PiAgentSession (the generic had one instantiation).
- Point handlers.ts at shared/agent/pi-events directly.
- Delete pi-runtime-state.ts.
